refactor: split probe detector into role-based submodules
