+++ src/pipeline/probe_detector.py
@@ -1,573 +1,41 @@
-"""Core probe scan detection logic moved from tools to src."""
+"""Core probe scan detection logic orchestrating submodules."""
 
 from __future__ import annotations
 
-import json
-from dataclasses import dataclass
 from pathlib import Path
-from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple
+from typing import Dict, List, Optional, Sequence, Tuple
 
 import cv2
 import numpy as np
 
-Box = Tuple[int, int, int, int]
+from src.pipeline.probe_detector_bands import (
+    build_divisi_map,
+    build_row_stats,
+    cluster_by_y_distance,
+    resolve_bands,
+    scan_staff_band_from_ink,
+    staff_bands_from_mask,
+)
+from src.pipeline.probe_detector_debug import write_debug_output
+from src.pipeline.probe_detector_rescue import apply_rightmost_rescue
+from src.pipeline.probe_detector_types import (
+    BandSelectionConfig,
+    Box,
+    DivisiRescueConfig,
+    RightmostRescueConfig,
+)
 
-_RIGHTMOST_RESCUE_DEBUG_KEYS = [
-    "band",
-    "staff_band",
-    "pred_band",
-    "ext_band",
-    "top_h",
-    "bottom_h",
-    "scan_band",
-    "scan_ext_band",
-    "scan_base_band",
-    "scan_row_ratio_mean",
-    "scan_row_ratio_max",
-    "scan_row_ratio_lines",
-    "scan_top_h",
-    "scan_bottom_h",
-    "scan_row_profile",
-    "scan_peak_ratio",
-    "scan_peak_row",
-    "scan_peak_ratio_local",
-    "scan_x_peak_ratio",
-    "scan_x_peak_neighbor_median",
-    "scan_x_peak_segment_min",
-    "scan_x_peak_segment_pass",
-    "scan_x_peak_ignored_rows",
+__all__ = [
+    "Box",
+    "BandSelectionConfig",
+    "DivisiRescueConfig",
+    "RightmostRescueConfig",
+    "cluster_by_y_distance",
+    "build_row_stats",
+    "staff_bands_from_mask",
+    "scan_staff_band_from_ink",
+    "detect_probe_scan",
 ]
-
-
-@dataclass(frozen=True)
-class BandSelectionConfig:
-    band_source: str
-    band_cluster_max_dist: float
-    band_min_row_count: int
-
-
-@dataclass(frozen=True)
-class DivisiRescueConfig:
-    enabled: bool
-    dist_ratio: float
-    align_tol: int
-    align_min_count: int
-    min_ratio: float
-    band_source: str
-    band_height_mode: str
-
-
-@dataclass(frozen=True)
-class RightmostRescueConfig:
-    enabled: bool
-    tolerance: int
-    min_rows: int
-    min_ratio: float
-
-
-def cluster_by_y_distance(y_centers: np.ndarray, max_distance: float, min_cluster_size: int):
-    sorted_indices = np.argsort(y_centers)
-    sorted_y = y_centers[sorted_indices]
-    clusters: List[List[int]] = []
-    current_cluster = [int(sorted_indices[0])]
-    for i in range(1, len(sorted_y)):
-        if sorted_y[i] - sorted_y[i - 1] <= max_distance:
-            current_cluster.append(int(sorted_indices[i]))
-        else:
-            clusters.append(current_cluster)
-            current_cluster = [int(sorted_indices[i])]
-    clusters.append(current_cluster)
-
-    valid_clusters: Dict[int, List[int]] = {}
-    noise: List[int] = []
-    cluster_id = 0
-    for cluster in clusters:
-        if len(cluster) >= min_cluster_size:
-            valid_clusters[cluster_id] = cluster
-            cluster_id += 1
-        else:
-            noise.extend(cluster)
-    return valid_clusters, noise
-
-
-def build_row_stats(
-    preds: Sequence[Box],
-    cluster_max_dist: float,
-    min_row_count: int,
-) -> List[Dict[str, float]]:
-    if not preds:
-        return []
-    y_centers = np.array([(box[1] + box[3]) / 2 for box in preds])
-    rows, _ = cluster_by_y_distance(y_centers, cluster_max_dist, min_row_count)
-    stats: List[Dict[str, float]] = []
-    for indices in rows.values():
-        if len(indices) < min_row_count:
-            continue
-        tops = [preds[i][1] for i in indices]
-        bottoms = [preds[i][3] for i in indices]
-        centers = [y_centers[i] for i in indices]
-        stats.append(
-            {
-                "center": float(np.median(centers)),
-                "top": float(np.median(tops)),
-                "bottom": float(np.median(bottoms)),
-            }
-        )
-    return stats
-
-
-def staff_bands_from_mask(
-    mask: np.ndarray,
-    gap_tolerance: int = 1,
-    min_height: int = 1,
-) -> List[Tuple[int, int]]:
-    rows = np.where(mask.sum(axis=1) > 0)[0]
-    if rows.size == 0:
-        return []
-    bands: List[Tuple[int, int]] = []
-    start = int(rows[0])
-    prev = int(rows[0])
-    for y in rows[1:]:
-        if int(y) - prev <= gap_tolerance:
-            prev = int(y)
-            continue
-        if prev - start + 1 >= min_height:
-            bands.append((start, prev))
-        start = int(y)
-        prev = int(y)
-    if prev - start + 1 >= min_height:
-        bands.append((start, prev))
-    return bands
-
-
-def scan_staff_band_from_ink(
-    ink: np.ndarray,
-    x_center: int,
-    y1: int,
-    y2: int,
-    scan_width: int,
-    line_ratio: float,
-    min_lines: int,
-    gap_tolerance: int = 1,
-) -> Tuple[int, int] | None:
-    """Estimate staff band by horizontal scan around x_center within [y1, y2]."""
-    h, w = ink.shape[:2]
-    x1 = max(0, int(x_center - scan_width // 2))
-    x2 = min(w - 1, int(x_center + scan_width // 2))
-    if x2 <= x1 or y2 <= y1:
-        return None
-    strip = ink[y1 : y2 + 1, x1 : x2 + 1]
-    if strip.size == 0:
-        return None
-    row_ratio = strip.sum(axis=1) / float(strip.shape[1])
-    rows = np.where(row_ratio >= line_ratio)[0]
-    if rows.size == 0:
-        return None
-    groups = []
-    start = int(rows[0])
-    prev = int(rows[0])
-    for r in rows[1:]:
-        if int(r) - prev <= gap_tolerance:
-            prev = int(r)
-            continue
-        groups.append((start, prev))
-        start = int(r)
-        prev = int(r)
-    groups.append((start, prev))
-    if len(groups) < min_lines:
-        return None
-    # Choose the tightest window of `min_lines` groups (staff lines),
-    # favoring higher ink ratios to avoid stray lines outside the staff.
-    best = None
-    for i in range(0, len(groups) - min_lines + 1):
-        j = i + min_lines - 1
-        span = groups[j][1] - groups[i][0]
-        mean_ratio = float(row_ratio[groups[i][0] : groups[j][1] + 1].mean())
-        if best is None or span < best[0] or (span == best[0] and mean_ratio > best[1]):
-            best = (span, mean_ratio, i, j)
-    if best is None:
-        return None
-    _, _, i, j = best
-    top = y1 + groups[i][0]
-    bottom = y1 + groups[j][1]
-    if bottom < top:
-        return None
-    return (int(top), int(bottom))
-
-
-def _resolve_bands(
-    *,
-    staff_mask: np.ndarray,
-    existing_boxes: Sequence[Box],
-    row_stats: Sequence[Dict[str, float]] | None,
-    config: BandSelectionConfig,
-) -> List[Tuple[int, int]]:
-    if config.band_source in ("existing_boxes", "horiz_scan", "row_stats"):
-        if config.band_source == "row_stats" and row_stats is not None:
-            return [
-                (int(stat["top"]), int(stat["bottom"]))
-                for stat in row_stats
-                if stat["bottom"] >= stat["top"]
-            ]
-        row_stats_local = build_row_stats(
-            existing_boxes, config.band_cluster_max_dist, config.band_min_row_count
-        )
-        return [
-            (int(stat["top"]), int(stat["bottom"]))
-            for stat in row_stats_local
-            if stat["bottom"] >= stat["top"]
-        ]
-    return staff_bands_from_mask(staff_mask)
-
-
-def _build_divisi_map(
-    *,
-    ink: np.ndarray,
-    bands: Sequence[Tuple[int, int]],
-    existing_boxes: Sequence[Box],
-    kernel: np.ndarray,
-    width: int,
-    image_h: int,
-    global_height: int,
-    config: DivisiRescueConfig,
-) -> Dict[int, Dict[str, bool]]:
-    divisi_map: Dict[int, Dict[str, bool]] = {}
-    if not config.enabled or len(bands) <= 1:
-        return divisi_map
-
-    adj_groups = []
-    b_heights = [y2 - y1 + 1 for y1, y2 in bands]
-    avg_h = float(np.mean(b_heights)) if b_heights else float(global_height)
-    dist_thresh = avg_h * config.dist_ratio
-
-    current_group = [0]
-    for i in range(1, len(bands)):
-        prev_y2 = bands[i - 1][1]
-        curr_y1 = bands[i][0]
-        dist = curr_y1 - prev_y2
-        if dist < dist_thresh:
-            current_group.append(i)
-        else:
-            if len(current_group) > 1:
-                adj_groups.append(current_group)
-            current_group = [i]
-    if len(current_group) > 1:
-        adj_groups.append(current_group)
-
-    band_xs: Dict[int, List[float]] = {i: [] for i in range(len(bands))}
-    for bx1, by1, bx2, by2 in existing_boxes:
-        cy = (by1 + by2) / 2.0
-        cx = (bx1 + bx2) / 2.0
-        best_bi = -1
-        best_dy = float("inf")
-        for i, (by1_b, by2_b) in enumerate(bands):
-            if by1_b <= cy <= by2_b:
-                band_xs[i].append(cx)
-                best_bi = -1
-                break
-            dy = min(abs(cy - by1_b), abs(cy - by2_b))
-            if dy < best_dy:
-                best_dy = dy
-                best_bi = i
-        if best_bi != -1:
-            band_xs[best_bi].append(cx)
-
-    for i, (y1, y2) in enumerate(bands):
-        band_center = int(round((y1 + y2) / 2))
-        band_h = max(1, y2 - y1 + 1)
-        target_h = band_h
-        if config.band_source == "row_stats":
-            band_y1, band_y2 = y1, y2
-        else:
-            if config.band_height_mode == "median_box":
-                target_h = band_h
-            band_y1 = max(0, int(band_center - target_h // 2))
-            band_y2 = min(image_h - 1, int(band_center + target_h // 2))
-
-        band_img = ink[band_y1 : band_y2 + 1, :]
-        if band_img.size == 0:
-            continue
-
-        col_sums = band_img.sum(axis=0)
-        stripe_sums = np.convolve(col_sums, kernel, mode="same")
-        ratios = stripe_sums / float(max(1, band_y2 - band_y1 + 1) * width)
-
-        divisi_min_ratio = config.min_ratio
-        if ratios.size < 3:
-            continue
-        peak_indices = np.where(
-            (ratios >= divisi_min_ratio)
-            & (ratios >= np.roll(ratios, 1))
-            & (ratios >= np.roll(ratios, -1))
-        )[0]
-        for px in peak_indices:
-            band_xs[i].append(float(px))
-
-    for grp in adj_groups:
-        for k in range(len(grp) - 1):
-            idx_a = grp[k]
-            idx_b = grp[k + 1]
-            xs_a = sorted(band_xs[idx_a])
-            xs_b = sorted(band_xs[idx_b])
-            match_count = 0
-            for xa in xs_a:
-                for xb in xs_b:
-                    if abs(xa - xb) <= config.align_tol:
-                        match_count += 1
-                        break
-            if match_count >= config.align_min_count:
-                if idx_a not in divisi_map:
-                    divisi_map[idx_a] = {"has_top": False, "has_bottom": False}
-                if idx_b not in divisi_map:
-                    divisi_map[idx_b] = {"has_top": False, "has_bottom": False}
-                divisi_map[idx_a]["has_bottom"] = True
-                divisi_map[idx_b]["has_top"] = True
-
-    return divisi_map
-
-
-def _apply_rightmost_rescue(
-    *,
-    config: RightmostRescueConfig,
-    accepted_by_band: Dict[int, List[float]],
-    trusted_accepted_by_band: Dict[int, List[float]],
-    rejected_records: List[Dict[str, Any]],
-    bands: Sequence[Tuple[int, int]],
-    has_existing: Callable[[float, int, int], bool],
-    candidates: List[Box],
-    debug_records: List[Dict[str, Any]],
-) -> None:
-    if not config.enabled or not accepted_by_band:
-        return
-
-    pool = trusted_accepted_by_band if trusted_accepted_by_band else accepted_by_band
-    rightmost_by_band_map = {band_idx: max(xs) for band_idx, xs in pool.items() if xs}
-    rightmost_values = list(rightmost_by_band_map.values())
-    if not rightmost_values:
-        return
-
-    max_col = max(rightmost_values)
-    if config.min_ratio > 0:
-        rightmost_values = [x for x in rightmost_values if x >= max_col * config.min_ratio]
-    if not rightmost_values:
-        rightmost_values = [max_col]
-    target = float(np.median(rightmost_values))
-
-    band_updates: dict[tuple[int, int], float] = {}
-    for rec in rejected_records:
-        band_as_list = rec["record"].get("staff_band")
-        if not band_as_list:
-            continue
-        col = float(rec["col"])
-        if abs(col - target) > config.tolerance:
-            continue
-        band_key = (int(band_as_list[0]), int(band_as_list[1]))
-        prev = band_updates.get(band_key)
-        if prev is None or abs(col - target) < abs(prev - target):
-            band_updates[band_key] = col
-
-    updated_values = []
-    for band_idx, col in rightmost_by_band_map.items():
-        band = bands[band_idx]
-        updated_values.append(band_updates.get((int(band[0]), int(band[1])), col))
-    inliers = [x for x in updated_values if abs(x - target) <= config.tolerance]
-    if len(inliers) < config.min_rows:
-        return
-
-    rescued_cols: set[int] = set()
-    rescue_statuses = {
-        "scan_ratio_low",
-        "scan_ratio_rel_low",
-        "extended_ratio_scan",
-        "extended_top_ratio_scan",
-        "extended_bottom_ratio_scan",
-    }
-    for rec in rejected_records:
-        rec_status = rec["record"].get("status")
-        if rec_status not in rescue_statuses:
-            continue
-        col = float(rec["col"])
-        if abs(col - target) > config.tolerance:
-            continue
-        col_i = int(round(col))
-        if col_i in rescued_cols:
-            continue
-        box = rec["box"]
-        if has_existing(col, box[1], box[3]):
-            continue
-        candidates.append(box)
-        rescued_cols.add(col_i)
-        debug_records.append(
-            {
-                "status": "rightmost_rescued",
-                "col": col,
-                "ratio": rec["record"].get("ratio"),
-                "extended_ratio": rec["record"].get("extended_ratio"),
-                "top_ratio": rec["record"].get("top_ratio"),
-                "bottom_ratio": rec["record"].get("bottom_ratio"),
-                "seed_col": rec["record"].get("seed_col"),
-                "rightmost_target": target,
-                "rightmost_delta": float(col - target),
-                "rightmost_band_updated": bool(band_updates),
-                **{k: rec["record"].get(k) for k in _RIGHTMOST_RESCUE_DEBUG_KEYS},
-            }
-        )
-
-
-def _draw_relative_band_rect(
-    crop: np.ndarray,
-    band: Any,
-    cy1: int,
-    cy2: int,
-    color: tuple[int, int, int],
-) -> None:
-    if not band:
-        return
-    y1, y2 = band
-    y1 = max(cy1, int(y1)) - cy1
-    y2 = min(cy2, int(y2)) - cy1
-    cv2.rectangle(crop, (0, y1), (crop.shape[1] - 1, y2), color, 1)
-
-
-def _draw_crop_band_overlays(crop: np.ndarray, rec: Dict[str, Any], cy1: int, cy2: int) -> None:
-    _draw_relative_band_rect(crop, rec.get("pred_band"), cy1, cy2, (0, 255, 0))
-    _draw_relative_band_rect(crop, rec.get("band"), cy1, cy2, (255, 0, 0))
-    _draw_relative_band_rect(crop, rec.get("ext_band"), cy1, cy2, (0, 0, 255))
-    _draw_relative_band_rect(crop, rec.get("scan_base_band"), cy1, cy2, (0, 255, 255))
-    _draw_relative_band_rect(crop, rec.get("scan_band"), cy1, cy2, (0, 165, 255))
-    _draw_relative_band_rect(crop, rec.get("scan_ext_band"), cy1, cy2, (128, 0, 255))
-
-
-def _draw_crop_labels(
-    crop: np.ndarray,
-    rec: Dict[str, Any],
-    *,
-    extend_top_max_ratio: float,
-    extend_bottom_max_ratio: float,
-) -> None:
-    ratio = rec.get("ratio")
-    top_ratio = rec.get("top_ratio")
-    bottom_ratio = rec.get("bottom_ratio")
-    ext_ratio = rec.get("extended_ratio")
-    scan_row_ratio_mean = rec.get("scan_row_ratio_mean")
-    scan_row_ratio_max = rec.get("scan_row_ratio_max")
-    scan_row_ratio_lines = rec.get("scan_row_ratio_lines")
-    scan_top_h = rec.get("scan_top_h")
-    scan_bottom_h = rec.get("scan_bottom_h")
-    label = (
-        f"{rec.get('status', '')} r={ratio:.2f} ext={ext_ratio:.2f}"
-        if ratio is not None and ext_ratio is not None
-        else rec.get("status", "")
-    )
-    cv2.putText(crop, label, (2, 14), cv2.FONT_HERSHEY_SIMPLEX, 0.4, (0, 0, 255), 1)
-    if top_ratio is not None:
-        cv2.putText(
-            crop,
-            f"top={top_ratio:.2f} <{extend_top_max_ratio:.2f}",
-            (2, 28),
-            cv2.FONT_HERSHEY_SIMPLEX,
-            0.4,
-            (255, 255, 0),
-            1,
-        )
-    if bottom_ratio is not None:
-        cv2.putText(
-            crop,
-            f"bot={bottom_ratio:.2f} <{extend_bottom_max_ratio:.2f}",
-            (2, 42),
-            cv2.FONT_HERSHEY_SIMPLEX,
-            0.4,
-            (255, 0, 255),
-            1,
-        )
-    if scan_row_ratio_mean is not None and scan_row_ratio_max is not None:
-        cv2.putText(
-            crop,
-            f"row_mean={scan_row_ratio_mean:.2f} row_max={scan_row_ratio_max:.2f} lines={scan_row_ratio_lines}",
-            (2, 56),
-            cv2.FONT_HERSHEY_SIMPLEX,
-            0.4,
-            (0, 0, 0),
-            1,
-        )
-    if scan_top_h is not None or scan_bottom_h is not None:
-        cv2.putText(
-            crop,
-            f"top_h={scan_top_h} bot_h={scan_bottom_h}",
-            (2, 70),
-            cv2.FONT_HERSHEY_SIMPLEX,
-            0.4,
-            (0, 0, 0),
-            1,
-        )
-
-
-def _write_debug_output(
-    *,
-    base_img: np.ndarray,
-    bands: Sequence[Tuple[int, int]],
-    debug_records: List[Dict[str, Any]],
-    debug_path: Path,
-    width: int,
-    params: Dict[str, Any],
-    divisi_map: Dict[int, Dict[str, bool]],
-    extend_top_max_ratio: float,
-    extend_bottom_max_ratio: float,
-) -> None:
-    h, w = base_img.shape[:2]
-    overlay = base_img.copy()
-    mask_overlay = overlay.copy()
-    for y1, y2 in bands:
-        cv2.rectangle(mask_overlay, (0, y1), (w - 1, y2), (255, 255, 0), -1)
-    overlay = cv2.addWeighted(mask_overlay, 0.2, overlay, 0.8, 0.0)
-    for rec in debug_records:
-        col = rec.get("col")
-        if col is None:
-            continue
-        color = (0, 255, 0) if rec["status"] == "accepted" else (0, 0, 255)
-        cv2.line(overlay, (int(col), 0), (int(col), h - 1), color, 1)
-    debug_path.parent.mkdir(parents=True, exist_ok=True)
-    cv2.imwrite(str(debug_path), overlay)
-    debug_json = debug_path.with_suffix(".json")
-    debug_json.write_text(
-        json.dumps(
-            {
-                "params": params,
-                "bands": bands,
-                "divisi_map": divisi_map,
-                "records": debug_records,
-            },
-            indent=2,
-        )
-    )
-    crop_dir = debug_path.parent / "endbar_debug_crops"
-    crop_dir.mkdir(parents=True, exist_ok=True)
-    for idx, rec in enumerate(debug_records):
-        col = rec.get("col")
-        if col is None:
-            continue
-        ext_band = rec.get("ext_band")
-        if ext_band:
-            cy1, cy2 = ext_band
-        else:
-            cy1, cy2 = rec.get("band", [0, h - 1])
-        cx1 = max(0, int(col) - width * 6)
-        cx2 = min(w - 1, int(col) + width * 6)
-        cy1 = max(0, int(cy1))
-        cy2 = min(h - 1, int(cy2))
-        crop = base_img[cy1 : cy2 + 1, cx1 : cx2 + 1].copy()
-        if crop.size == 0:
-            continue
-        _draw_crop_band_overlays(crop, rec, cy1, cy2)
-        cv2.line(crop, (int(col - cx1), 0), (int(col - cx1), crop.shape[0] - 1), (0, 0, 255), 1)
-        _draw_crop_labels(
-            crop,
-            rec,
-            extend_top_max_ratio=extend_top_max_ratio,
-            extend_bottom_max_ratio=extend_bottom_max_ratio,
-        )
-        name = f"{idx:04d}_{rec.get('status', 'status')}_col{col}.png"
-        cv2.imwrite(str(crop_dir / name), crop)
 
 
 def detect_probe_scan(
@@ -645,7 +113,7 @@ def detect_probe_scan(
         band_cluster_max_dist=band_cluster_max_dist,
         band_min_row_count=band_min_row_count,
     )
-    bands = _resolve_bands(
+    bands = resolve_bands(
         staff_mask=staff_mask,
         existing_boxes=existing_boxes,
         row_stats=row_stats,
@@ -668,7 +136,7 @@ def detect_probe_scan(
         band_source=band_source,
         band_height_mode=band_height_mode,
     )
-    divisi_map = _build_divisi_map(
+    divisi_map = build_divisi_map(
         ink=ink,
         bands=bands,
         existing_boxes=existing_boxes,
@@ -1348,7 +816,7 @@ def detect_probe_scan(
         min_rows=scan_rightmost_min_rows,
         min_ratio=scan_rightmost_min_ratio,
     )
-    _apply_rightmost_rescue(
+    apply_rightmost_rescue(
         config=rightmost_cfg,
         accepted_by_band=accepted_by_band,
         trusted_accepted_by_band=trusted_accepted_by_band,
@@ -1407,7 +875,7 @@ def detect_probe_scan(
             "scan_ratio_rel_rescue_xpeak_min": scan_ratio_rel_rescue_xpeak_min,
             "scan_ratio_rel_rescue_max_overhang": scan_ratio_rel_rescue_max_overhang,
         }
-        _write_debug_output(
+        write_debug_output(
             base_img=base_img,
             bands=bands,
             debug_records=debug_records,
